requirements.txt
diffusers==0.16.1
numpy==1.24.3
Pillow==9.5.0
Pillow==9.5.0
torch==2.0.0
